scripts/requirements.txt
yt-dlp>=2024.0
openai>=1.0
anthropic>=0.40
python-dotenv
tqdm
aiohttp
aiofiles
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pgvector>=0.3
